IOT-44: Authorization on IoTDevice CRUD

+++ src/controllers/admin-controller/iot-device.controller.ts
@@ -10,12 +10,17 @@ import {
     Delete,
     BadRequestException,
     Logger,
+    UseGuards,
+    Req,
 } from "@nestjs/common";
 import {
     ApiTags,
     ApiOperation,
     ApiBadRequestResponse,
     ApiNotFoundResponse,
+    ApiBearerAuth,
+    ApiForbiddenResponse,
+    ApiUnauthorizedResponse,
 } from "@nestjs/swagger";
 import { IoTDeviceService } from "@services/iot-device.service";
 import { CreateIoTDeviceDto } from "@dto/create-iot-device.dto";
@@ -23,16 +28,33 @@ import { IoTDevice } from "@entities/iot-device.entity";
 import { UpdateIoTDeviceDto } from "@dto/update-iot-device.dto";
 import { DeleteResponseDto } from "@dto/delete-application-response.dto";
 import { ErrorCodes } from "@enum/error-codes.enum";
+import {
+    checkIfUserHasReadAccessToApplication,
+    checkIfUserHasWriteAccessToApplication,
+} from "@helpers/security-helper";
+import { AuthenticatedRequest } from "@dto/internal/authenticated-request";
+import { JwtAuthGuard } from "@auth/jwt-auth.guard";
+import { RolesGuard } from "@auth/roles.guard";
+import { Read } from "@auth/roles.decorator";
+import { response } from "express";
 
 @ApiTags("IoT Device")
 @Controller("iot-device")
+@UseGuards(JwtAuthGuard, RolesGuard)
+@ApiBearerAuth()
+@Read()
+@ApiForbiddenResponse()
+@ApiUnauthorizedResponse()
 export class IoTDeviceController {
     constructor(private iotDeviceService: IoTDeviceService) {}
 
     @Get(":id")
     @ApiOperation({ summary: "Find one IoT-Device by id" })
     @ApiNotFoundResponse()
-    async findOne(@Param("id") id: number): Promise<IoTDevice> {
+    async findOne(
+        @Req() req: AuthenticatedRequest,
+        @Param("id") id: number
+    ): Promise<IoTDevice> {
         let result = undefined;
         try {
             result = await this.iotDeviceService.findOneWithApplicationAndMetadata(
@@ -47,6 +69,9 @@ export class IoTDeviceController {
         if (!result) {
             throw new NotFoundException(ErrorCodes.IdDoesNotExists);
         }
+
+        checkIfUserHasReadAccessToApplication(req, result.application.id);
+
         return result;
     }
 
@@ -54,7 +79,12 @@ export class IoTDeviceController {
     @Header("Cache-Control", "none")
     @ApiOperation({ summary: "Create a new IoTDevice" })
     @ApiBadRequestResponse()
-    async create(@Body() createDto: CreateIoTDeviceDto): Promise<IoTDevice> {
+    async create(
+        @Req() req: AuthenticatedRequest,
+        @Body() createDto: CreateIoTDeviceDto
+    ): Promise<IoTDevice> {
+        checkIfUserHasWriteAccessToApplication(req, createDto.applicationId);
+
         const application = this.iotDeviceService.create(createDto);
         return application;
     }
@@ -64,9 +94,26 @@ export class IoTDeviceController {
     @ApiOperation({ summary: "Update an existing IoT-Device" })
     @ApiBadRequestResponse()
     async update(
+        @Req() req: AuthenticatedRequest,
         @Param("id") id: number,
         @Body() updateDto: UpdateIoTDeviceDto
     ): Promise<IoTDevice> {
+        // Old application
+        const oldIotDevice = await this.iotDeviceService.findOneWithApplicationAndMetadata(
+            id
+        );
+        checkIfUserHasWriteAccessToApplication(
+            req,
+            oldIotDevice.application.id
+        );
+        if (updateDto.applicationId !== oldIotDevice.application.id) {
+            // New application
+            checkIfUserHasWriteAccessToApplication(
+                req,
+                updateDto.applicationId
+            );
+        }
+
         const application = await this.iotDeviceService.update(id, updateDto);
 
         return application;
@@ -75,7 +122,18 @@ export class IoTDeviceController {
     @Delete(":id")
     @ApiOperation({ summary: "Delete an existing IoT-Device" })
     @ApiBadRequestResponse()
-    async delete(@Param("id") id: number): Promise<DeleteResponseDto> {
+    async delete(
+        @Req() req: AuthenticatedRequest,
+        @Param("id") id: number
+    ): Promise<DeleteResponseDto> {
+        const oldIotDevice = await this.iotDeviceService.findOneWithApplicationAndMetadata(
+            id
+        );
+        checkIfUserHasWriteAccessToApplication(
+            req,
+            oldIotDevice?.application?.id
+        );
+
         try {
             const result = await this.iotDeviceService.delete(id);
             return new DeleteResponseDto(result.affected);
